[+] Added 10 products (on bootstrap) [~] Changed datatype of `Product` from `Double` to `int`

+++ webshop/src/main/java/com/lowtech/webshop/model/Product.java
@@ -16,7 +16,10 @@ public class Product extends AbstractEntity {
 
     private String name;
     private String description;
-    private Double price;
+
+    // cents
+    private int price;
+
     private String category;
     private String imageUrl;
 }
